Make EnterAddress as shared component
