+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/sendtransaction/AbstractSendTransactionService.kt
@@ -28,6 +28,8 @@ abstract class AbstractSendTransactionService(val hasSettings: Boolean, val hasN
     open fun GetSettingsContent(navController: NavController) = Unit
     @Composable
     open fun GetNonceSettingsContent(navController: NavController) = Unit
+    @Composable
+    open fun GetRecipientSettingsContent(navController: NavController) = Unit
     abstract suspend fun sendTransaction(mevProtectionEnabled: Boolean = false): SendTransactionResult
 
     fun refreshUuid() {
