+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/sendtransaction/SendTransactionServiceEvm.kt
@@ -1,10 +1,19 @@
 package io.horizontalsystems.bankwallet.modules.multiswap.sendtransaction
 
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.ime
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
@@ -16,7 +25,10 @@ import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.ethereum.CautionViewItem
 import io.horizontalsystems.bankwallet.core.ethereum.CautionViewItemFactory
 import io.horizontalsystems.bankwallet.core.ethereum.EvmCoinServiceFactory
+import io.horizontalsystems.bankwallet.entities.Address
 import io.horizontalsystems.bankwallet.entities.DataState
+import io.horizontalsystems.bankwallet.modules.enteraddress.EnterAddress
+import io.horizontalsystems.bankwallet.modules.evmfee.ButtonsGroupWithShade
 import io.horizontalsystems.bankwallet.modules.evmfee.Cautions
 import io.horizontalsystems.bankwallet.modules.evmfee.Eip1559FeeSettings
 import io.horizontalsystems.bankwallet.modules.evmfee.EvmCommonGasDataService
@@ -39,6 +51,7 @@ import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmSettings
 import io.horizontalsystems.bankwallet.modules.send.evm.settings.SendEvmSettingsViewModel
 import io.horizontalsystems.bankwallet.ui.compose.ComposeAppTheme
 import io.horizontalsystems.bankwallet.ui.compose.TranslatableString
+import io.horizontalsystems.bankwallet.ui.compose.components.ButtonPrimaryYellow
 import io.horizontalsystems.bankwallet.ui.compose.components.MenuItem
 import io.horizontalsystems.bankwallet.ui.compose.components.VSpacer
 import io.horizontalsystems.bankwallet.uiv3.components.HSScaffold
@@ -49,6 +62,7 @@ import io.horizontalsystems.ethereumkit.decorations.TransactionDecoration
 import io.horizontalsystems.ethereumkit.models.GasPrice
 import io.horizontalsystems.ethereumkit.models.TransactionData
 import io.horizontalsystems.marketkit.models.BlockchainType
+import io.horizontalsystems.marketkit.models.Token
 import io.reactivex.Flowable
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -260,6 +274,14 @@ class SendTransactionServiceEvm(
     }
 
     @Composable
+    override fun GetRecipientSettingsContent(navController: NavController) {
+        SendEvmRecipientSettingsScreen(
+            token = token,
+            navController = navController,
+        )
+    }
+
+    @Composable
     override fun GetSettingsContent(navController: NavController) {
         val feeSettingsViewModel = viewModel<ViewModel>(
             factory = EvmFeeModule.Factory(
@@ -366,6 +388,64 @@ fun SendEvmNonceSettingsScreen(
             Cautions(viewModel.cautions)
 
             VSpacer(32.dp)
+        }
+    }
+}
+
+@Composable
+fun SendEvmRecipientSettingsScreen(
+    token: Token,
+    navController: NavController,
+) {
+    var hasValidationError by remember { mutableStateOf(false) }
+    var address by remember { mutableStateOf<Address?>(null) }
+    var validationInProgress by remember { mutableStateOf(false) }
+
+    HSScaffold(
+        title = stringResource(R.string.SendEvmSettings_SetRecipient),
+        onBack = navController::popBackStack,
+    ) {
+        Column(
+            modifier = Modifier.windowInsetsPadding(WindowInsets.ime)
+        ) {
+            Column(
+                modifier = Modifier
+                    .verticalScroll(rememberScrollState())
+                    .weight(1f)
+                    .fillMaxSize()
+            ) {
+                EnterAddress(
+                    modifier = Modifier,
+                    navController = navController,
+                    token = token,
+                    onValueChange = {
+                        address = it
+                    },
+                    onValidationError = { error ->
+                        hasValidationError = error != null
+                    },
+                    onValidationInProgress = {
+                        validationInProgress = it
+                    }
+                )
+
+                VSpacer(32.dp)
+            }
+            ButtonsGroupWithShade {
+                ButtonPrimaryYellow(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(start = 16.dp, end = 16.dp),
+                    title = if (hasValidationError)
+                        stringResource(R.string.Send_Address_Error_InvalidAddress)
+                    else
+                        stringResource(R.string.Button_Next),
+                    onClick = {
+                        navController.popBackStack()
+                    },
+                    enabled = address != null && !validationInProgress && !hasValidationError
+                )
+            }
         }
     }
 }
