+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/settings/SwapTransactionRecipientSettingsFragment.kt
@@ -1,0 +1,25 @@
+package io.horizontalsystems.bankwallet.modules.multiswap.settings
+
+import androidx.compose.runtime.Composable
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavController
+import io.horizontalsystems.bankwallet.core.BaseComposeFragment
+import io.horizontalsystems.bankwallet.modules.multiswap.SwapConfirmViewModel
+
+class SwapTransactionRecipientSettingsFragment : BaseComposeFragment() {
+    @Composable
+    override fun GetContent(navController: NavController) {
+        SwapRecipientSettingsScreen(navController)
+    }
+}
+
+@Composable
+fun SwapRecipientSettingsScreen(navController: NavController) {
+    val viewModel = viewModel<SwapConfirmViewModel>(
+        viewModelStoreOwner = navController.previousBackStackEntry!!,
+    )
+
+    val sendTransactionService = viewModel.sendTransactionService
+
+    sendTransactionService.GetRecipientSettingsContent(navController)
+}
