+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/multiswap/settings/ui/SwapSettingsComponents.kt
@@ -105,7 +105,6 @@ fun RecipientAddress(
         initial = initial,
         tokenQuery = token.tokenQuery,
         coinCode = token.coin.code,
-        navController = navController,
         onError = onError,
         onValueChange = onValueChange,
     )
