+++ app/src/main/res/navigation/main_graph.xml
@@ -325,6 +325,9 @@
         android:id="@+id/swapTransactionNonceSettings"
         android:name="io.horizontalsystems.bankwallet.modules.multiswap.settings.SwapTransactionNonceSettingsFragment" />
     <fragment
+        android:id="@+id/swapTransactionRecipientSettings"
+        android:name="io.horizontalsystems.bankwallet.modules.multiswap.settings.SwapTransactionRecipientSettingsFragment" />
+    <fragment
         android:id="@+id/swapConfirm"
         android:name="io.horizontalsystems.bankwallet.modules.multiswap.SwapConfirmFragment" />
     <fragment
